// src/enums/card-type.ts
export enum CardType {
    CHALLENGE = 'challenge',
    COMMUNICATION = 'communication',
    CONTRIBUTION = 'contribution',
    CREATIVENESS = 'creativeness',
    EXPERT = 'expert',
    FEEDBACK = 'feedback',
    LEADER = 'leader',
    LIFESTYLE = 'lifestyle',
    SALARY = 'salary',
    STATUS = 'status',
    TEAM = 'team',
}